app/requirements.txt
tweepy==4.4.0
python-dotenv==0.19.2
psycopg2-binary==2.9.2